Major bug fixes and preliminary work.
Task was no completed due to the need to fix maxjor issues that have existed
since the beginning of this project. The messaging system has been having
problems with correctly displaying information that it receieves from the
server. This pre-commit was an attempt to solve that. Further commits are
necessary.