Added new statistics table and display statistics
A new statistics table for the phpmyadmin database has been added into the create_tables.sql file. Page visits for index, messages and statistics are now recorded and stored in the new table as well as displayed on the statistics page. Also fixed minor bug with getting MySql permissions and '%' host users.