Database updates now working correctly.
Issue was with how the update function was being called. Previously was being
called any time the page was loaded. This meant that AJAX calls would trigger
the database to be updated. Since the the Message page JS script always calls
an AJAX query as soon as the page loads, the database would be updated twice
for every page load.